nevermind, I was just doing something idiotic. it works well now, not perfect, but pretty well

+++ src/Player.h
@@ -45,7 +45,7 @@ private:
     Controls controls;
     ofRay playerRay;
     ofPlane intersectionPlane;
-    vector<pair<int, int> > distances;
+    vector<pair<float, int> > distances;
     float safeY;
 };
